+++ src/Recyclarr.TrashLib/Compatibility/Sonarr/ISonarrCapabilityFetcher.cs
@@ -2,7 +2,7 @@ using Recyclarr.TrashLib.Config.Services;
 
 namespace Recyclarr.TrashLib.Compatibility.Sonarr;
 
-public interface ISonarrCapabilityChecker
+public interface ISonarrCapabilityFetcher
 {
     Task<SonarrCapabilities?> GetCapabilities(IServiceConfiguration config);
 }