+++ src/Recyclarr.TrashLib/Compatibility/Sonarr/SonarrCapabilityFetcher.cs
@@ -2,17 +2,19 @@ using Recyclarr.TrashLib.ApiServices.System;
 
 namespace Recyclarr.TrashLib.Compatibility.Sonarr;
 
-public class SonarrCapabilityChecker : ServiceCapabilityChecker<SonarrCapabilities>, ISonarrCapabilityChecker
+public class SonarrCapabilityFetcher : ServiceCapabilityFetcher<SonarrCapabilities>, ISonarrCapabilityFetcher
 {
-    public SonarrCapabilityChecker(IServiceInformation info)
+    public SonarrCapabilityFetcher(IServiceInformation info)
         : base(info)
     {
     }
 
     protected override SonarrCapabilities BuildCapabilitiesObject(Version version)
     {
-        return new SonarrCapabilities(version)
+        return new SonarrCapabilities
         {
+            Version = version,
+
             SupportsNamedReleaseProfiles =
                 version >= SonarrCapabilities.MinimumVersion,