+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- When using `sync`, continue processing other instances when there's a failure.
+
 ## [5.0.2] - 2023-06-24
 
 ### Fixed