+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - When using `sync`, continue processing other instances when there's a failure.
+- Regression: Perform Sonarr compatibility checks again (#189).
 
 ## [5.0.2] - 2023-06-24